renamed js folder as common + moved hardcoded colors into global.js
