+++ src/components/Layout.js
@@ -16,7 +16,7 @@ import Notification from './Notification.jsx'
 import StaticPage from './StaticPage'
 import TemplateCover from './presentational/covers/TemplateCover'
 
-import { parseDate } from '../js/utilities'
+import { parseDate } from '../common/utilities'
 import { isMobile } from 'react-device-detect'
 
 class Dashboard extends React.Component {
