+++ src/components/Overlay/Loading.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import copy from '../../js/data/copy.json'
+import copy from '../../common/data/copy.json'
 
 const LoadingOverlay = ({ isLoading, language }) => {
   let classes = 'loading-overlay'
