+++ src/components/Timeline.jsx
@@ -4,8 +4,8 @@ import { connect } from 'react-redux'
 import * as selectors from '../selectors'
 import hash from 'object-hash'
 
-import copy from '../js/data/copy.json'
-import { formatterWithYear, parseDate } from '../js/utilities'
+import copy from '../common/data/copy.json'
+import { formatterWithYear, parseDate } from '../common/utilities'
 import Header from './presentational/Timeline/Header'
 import Axis from './TimelineAxis.jsx'
 import Clip from './presentational/Timeline/Clip'
