+++ src/components/Toolbar/Search.js
@@ -1,6 +1,6 @@
 /* global fetch */
 import React from 'react'
-import copy from '../../js/data/copy.json'
+import copy from '../../common/data/copy.json'
 import TagFilter from './TagFilter'
 
 export default class Search extends React.Component {
