+++ src/components/presentational/Card/Category.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { capitalizeFirstLetter } from '../../../js/utilities.js'
+import { capitalizeFirstLetter } from '../../../common/utilities.js'
 
 const CardCategory = ({ categoryTitle, categoryLabel, color }) => (
   <div className='card-row card-cell category'>
