+++ src/components/presentational/Card/Location.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import copy from '../../../js/data/copy.json'
+import copy from '../../../common/data/copy.json'
 
 const CardLocation = ({ language, location, isPrecise }) => {
   if (location !== '') {
