+++ src/components/presentational/Card/Summary.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import copy from '../../../js/data/copy.json'
+import copy from '../../../common/data/copy.json'
 
 const CardSummary = ({ language, description, isHighlighted }) => {
   const summary = copy[language].cardstack.description
