+++ src/components/presentational/Card/Tags.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import copy from '../../../js/data/copy.json'
+import copy from '../../../common/data/copy.json'
 
 const CardTags = ({ tags, language }) => {
   const tagsLang = copy[language].cardstack.tags
